Add earlier work on timezones.
Add earlier work on timezones, with a few inline modifications. This is
really to big of a commit. But we are so far from a stable release that
it should be fine.

The current version seems to eager, and recalculates to many times. This
will soon be fixed in a future version.