Add timezone handling.
This is a way to large commit. But I see no feasable way to break it
down into multiple smaller commits.

The main "secret" to solving timezones for recurring events was to
remember to recalculate timezones whenever a new instance of the object
was generated.

This current implementation seems really slow (> 1s). Further testing is
needed.