Add flags for just running the text flower.
It's twisted if one program should do many things, or just a single
thing. If we assume that this program does two things: reading calendar
files, and writing output, then it should be ok to allow any of those
parts to run independentally of the other. Making it ok to just run the
text flower.