Rework program initialization.
Old init setup had the fancy idea to parse all files before anything
could be done with them. This however led to problems when a part of the
program which didn't care for the calendar files (such as text
formatting).

It also made testing performance almost impossible since to much code
was run before I had a chance to init statprof.