ICAL handling of events different from display handling.
Previously repeating events where always instantiated to a stream of all
events to come (possibly infinite), and then zipped with the list of
regular events to create a stream of all events in the world.

This commit allows access to the raw lists of parsed regular and
repeating events before they are extrapolated and merged.