HTML output only describe multi-day events once.
The HTML renderer mostly works in days. Events spanning over midnight
therefore appears multiple times in the layout. This lead to their
start-time + uid not to be unique any more, which was bad since I
assumed it to be unique.

This commit only describes a multi-day event on the day the event
starts, with a special case for events which start before a given time
span.

Repeating events aren't affected, and are still show once per instance.