Resolve (datetime instance) TODO with ./configure?
Tried adding a ./configure script, which mostly is responsible for
downloading a default zoneinfo file, and setting up the environment for
the program.

I have for quite a while thought about having a configure system for
things like these, but also for setting up default paths. Let's see if
it works out.