Refactor read-file to use string ports.
Like the diff mentions; appending strings leads to a huge number of
allocations and deallocations, killing performance in the GC for large
strings.

String ports on the other hand say nothing about how the work
internally, allowing Guile to have efficient C code which completely
bypasses the GC (this was briefly checked in Guile's source code, commit
118ee0c50ba90ea7ad7ff1fd2a212bbbb7a66d99 (3.0.7 + a few commits).

The approximate speed increase is at least two orders of magnitude, if
not three.