Allow partial change and removal of XML Namespaces.
Changing the error to a returnu allows a partial namespace change map,
which can sometimes be usefull (and should hopefully not lead to any
issues here).

Changing assoc-ref to assoc allows the removal of namespaces, since
assoc-ref can't differentiate between a match with #f in its cdr, and no
match.

    (assoc-ref '((a . #f)) 'a) ⇒ #f
    (assoc-ref '((a . #f)) 'b) ⇒ #f