Remove with-stream macro.
It was an extra (rather complicated) macro to support, which had 2
unresolved TODO:s, and made the resulting code hard to read since it
wasn't immediately apparent which procedures where replaced by their
stream counterparts.

The only using code was rewritten using a threading macro, which is way
more readable.