Change printer for vcomponent.
This new printer attempts to output something actually matching a
structure which the object can be recreated from. It's still not
read-compatible (since that can only be done through reader macros),
but a read + eval would work.

It's fed through the pretty printer, both so the output might actually
be readable, but also so that line diffs work (and so that acidental
output of the entire database doesn't crash emacs).

This still needs to be extended to the vlines to be really useful.