Change where and how virtual properties are tracked.
Previously, virtual properties where implemented on the root resource.
This moves it to the virtual resource type, since that's where they
semantically belong (seeing as the base property table was ignored by
the file resource type).

It also slightly changes the api for setting resources, renaming all the
procedures which returns promises to have a `!!` suffix.