Change iCalendar parser.
This splits the iCalendar parser into multiple modules, making it
clearer when we if we are looking at the structure of the file, or the
contents of the properties.

It also rewrites the line-unfolding algorithm to work on bytes instead
of characters, allowing badly formed streams with split multi-byte
characters to be properly parsed.