Add entry point for listing events.
This entry point is mostly a proof-of-concept that the new object system
can be used. It's however important for debugging purposes, and will
eventually be extended to a basic CLI interface. The main reason for it
being unusable as a CLI interface currently is due to the fact that it
needs to reload all databases on every invocation.